# contracts/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Middagsklok.Api",
    "version": "1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Middagsklok.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/dishes": {
      "get": {
        "tags": [
          "Middagsklok.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/dishes/import": {
      "post": {
        "tags": [
          "Middagsklok.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchImportDishesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/weekly-plan/{weekStartDate}": {
      "get": {
        "tags": [
          "Middagsklok.Api"
        ],
        "parameters": [
          {
            "name": "weekStartDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/weekly-plan/generate": {
      "post": {
        "tags": [
          "Middagsklok.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateWeeklyPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BatchImportDishesRequest": {
        "type": "object",
        "properties": {
          "dishes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DishImportItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DishImportItem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "activeMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "totalMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "kidRating": {
            "type": "integer",
            "format": "int32"
          },
          "familyRating": {
            "type": "integer",
            "format": "int32"
          },
          "isPescetarian": {
            "type": "boolean"
          },
          "hasOptionalMeatVariant": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DishIngredientItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DishIngredientItem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "optional": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DishListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "activeMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "totalMinutes": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DishListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DishListItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GenerateWeeklyPlanRequest": {
        "type": "object",
        "properties": {
          "weekStartDate": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
  "tags": [
    {
      "name": "Middagsklok.Api"
    }
  ]
}